Move permissions from models folder
