+++ synapseclient/__init__.py
@@ -29,7 +29,7 @@ from .table import (
 )
 from .team import Team, UserProfile, UserGroupHeader, TeamMember
 from .wiki import Wiki
-from synapseclient.models.permission import Permissions
+from .core.models.permission import Permissions
 
 # ref = importlib.resources.files(__name__).joinpath("synapsePythonClient")
 # with ref.open("r") as fp:
