+++ synapseclient/client.py
@@ -119,7 +119,7 @@ from synapseclient.core.upload.upload_functions import (
 from synapseclient.core.dozer import doze
 
 from typing import Union, Dict, List, Optional, Tuple
-from synapseclient.models.permission import Permissions
+from synapseclient.core.models.permission import Permissions
 from opentelemetry import trace
 
 tracer = trace.get_tracer("synapseclient")
