+++ synapseclient/core/models/permission.py
