--- synapseclient/models/__init__.py
@@ -1,4 +1,0 @@
-# These are all of the models that are used by the Synapse client.
-from .permission import Permissions
-
-__all__ = ["Permissions"]
